@@ -1,13 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Consul" Version="1.6.1.1" />
-    <PackageReference Include="Microsoft.ReverseProxy" Version="1.0.0-preview.4.20424.2" />
-    <PackageReference Include="System.Interactive" Version="4.1.1" />
+    <PackageReference Include="Consul" Version="1.7.14.1" />
+    <PackageReference Include="System.Interactive" Version="6.0.1" />
+    <PackageReference Include="Yarp.ReverseProxy" Version="2.1.0" />
   </ItemGroup>
 
 </Project>
