@@ -1,34 +1,34 @@
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Bogus;
 
-namespace ItemsApi
+namespace ItemsApi;
+
+public static class DataService
 {
-    public static class DataService
+    private static List<Item> Items { get; }
+
+    static DataService()
     {
-        private static List<Item> Items { get; set; }
+        Items = new Faker<Item>()
+            .CustomInstantiator( f=>
+                new Item(
+                    f.IndexVariable++,
+                    f.Commerce.ProductName(),
+                    f.PickRandom(f.Commerce.Categories(5)),
+                    f.Address.Country(),
+                    f.Random.Number(10, 100) )
+            ).Generate(50);
+    }
 
-        static DataService()
-        {
-            Items = new Faker<Item>()
-                .RuleFor(i => i.ID, f => f.IndexVariable++)
-                .RuleFor(i => i.Name, f => f.Commerce.ProductName())
-                .RuleFor(i => i.Category, f => f.PickRandom(f.Commerce.Categories(5)))
-                .RuleFor(i => i.Origin, f => f.Address.Country())
-                .RuleFor(i => i.Quantity, f => f.Random.Number(10, 100))
-                .Generate(50);
-        }
+    public static ValueTask<IEnumerable<Item>> GetItems()
+    {
+        return new ValueTask<IEnumerable<Item>>(Items);
+    }
 
-        public static ValueTask<IEnumerable<Item>> GetItems()
-        {
-            return new ValueTask<IEnumerable<Item>>(Items);
-        }
-
-        public static ValueTask<Item> GetItemByID(int id)
-        {
-            var item = Items.SingleOrDefault(i => i.ID == id);
-            return item != null ? new ValueTask<Item>(item) : new ValueTask<Item>();
-        }
+    public static ValueTask<Item?> GetItemById(int id)
+    {
+        var item = Items.SingleOrDefault(i => i.Id == id);
+        return new ValueTask<Item?>(item);
     }
 }
+
+public record Item(int Id, string Name, string Category, string Origin, int Quantity);
