@@ -1,48 +1,36 @@
-﻿using System;
-using System.Linq;
-using System.Net;
+﻿using System.Net.Http.Headers;
 using Consul;
-using ItemsApi.Workers;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
-namespace ItemsApi
+namespace ItemsApi;
+
+public static class ServiceCollectionExtensions
 {
-    public static class ServiceCollectionExtensions
+    public static IServiceCollection AddConsulClient(this IServiceCollection services,
+        IConfigurationSection config)
     {
-        public static IServiceCollection AddConsulClient(this IServiceCollection services)
+        services.AddHttpClient("Consul", client =>
         {
-            return services.AddConsulClient(options => { });
-        }
+            client.DefaultRequestHeaders.Accept.Add(
+                new MediaTypeWithQualityHeaderValue("application/json"));
+        });
 
-        public static IServiceCollection AddConsulClient(
-            this IServiceCollection services,
-            Action<ConsulClientConfiguration> options)
+        var host = config.GetValue<string>("host") ?? string.Empty;
+        var dc = config.GetValue<string>("datacenter") ?? string.Empty;
+
+        var consulClientConfiguration = new ConsulClientConfiguration
         {
-            /*
-             * CONSUL_HTTP_ADDR
-             * CONSUL_HTTP_SSL
-             * CONSUL_HTTP_SSL_VERIFY
-             * CONSUL_HTTP_AUTH
-             * CONSUL_HTTP_TOKEN
-             */
-            services.TryAddSingleton<IConsulClient>(sp => new ConsulClient(options));
+            Address = new Uri(host),
+            Datacenter = dc
+        };
 
-            return services;
-        }
-
-        public static IServiceCollection AddConsulServiceRegistration(this IServiceCollection services,
-            Action<AgentServiceRegistration> options)
+        services.TryAddTransient<IConsulClient>(sp =>
         {
-            return services
-                .AddSingleton(provider =>
-                {
-                    var registration = new AgentServiceRegistration();
-                    options.Invoke(registration);
+            var clientFactory = sp.GetRequiredService<IHttpClientFactory>();
 
-                    return registration;
-                })
-                .AddHostedService<ConsulRegistrationService>();
-        }
+            return new ConsulClient(consulClientConfiguration,
+                clientFactory.CreateClient("Consul"));
+        });
+        return services;
     }
 }
